dma-heap: qcom: Split sgl_desc into 2 separate variables
It is possible for sgl_desc to be freed by mem_buf_free() a couple of
lines above before kfree() is called on it in the error handling path
which can lead to a double free.

Since sgl_desc is allocated twice for two different cases within the
same function, separate it into two distinct variables which can be
used for each of the usecases. This makes handling the alloc/free
operations cleaner and also avoids the double free case.

Change-Id: Ie80ba8f39046e2c803280627d4830ef62d92b861
Signed-off-by: Oreoluwa Babatunde <quic_obabatun@quicinc.com>