sched/cpufreq_walt: Evaluate hispeed and nl only at window rollover
Hispeed condition is entered when the current util of a core reaches a
percentage of the cluster's average capacity. The average capacity of a
cluster is calculated as the util corresponding to the average frequency
seen by the cluster in the past window. At every frequency update, the
number of CPU cycles are calculated based on the previous frequency and
the duration that frequency lasted. These are accumulated in a counter.
The total number of cycles are then divided by the window length to get
the average frequency seen in the window.

Waltgov callbacks may occur multiple times within a window and the average
capacity will be the same as at the start of the window while the CPU's
util may have increased. This results in an incorrect comparison of the
previous window's average capacity with the current window's util when
hispeed is checked at each of these waltgov callbacks. Hispeed
evaluation should be skipped in these situations.

Currently, only migration induced waltgov callback skips checking for
hispeed. Extend it such that all reasons for waltgov callback except
window rollover skip checking for hispeed.

Since new workload detection is directly dependent on hispeed condition,
ensure that the check for it is also skipped for all reasons except
window rollover.

Change-Id: I658921b2e6110a8b867d4b1db93d64e428835f92
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>