walt: cfs: fix cpu_util_without in overloaded condition
cpu_util_without() function is used to remove a task's utilization from
a cpu's utilization if the task is running or enqueued on that cpu. This
is essential during cpu selection for a task that is already running on
a cpu.

However, the current implementation is broken because cpu_util() caps
the cpu_util at its capacity.

For example, assume a cpu with capacity of 1024 units is packed with tow
tasks of 500 units and a task of 600 units.  If the 600unit task is
requested for cpu_util_without(), the return value should be 1000
= (600 + 500 + 500) - 600. But since cpu_util() caps to 1024 the return
value is 1024 - 600 = 424.

This returns a lower value for utilization from other tasks which causes
a higher than actual spare capacity in the placement path. This was seen
to cause more packing on the already packed cpu.

Fix this by using an uncapped util, cumulative_runnable_avg_scaled,
instead of cpu_util() for subtraction.

Change-Id: I3be805913cecc5d3c84712ca0571a4fb18213219
Signed-off-by: Abhijeet Dharmapurikar <quic_adharmap@quicinc.com>