sched/walt: Add calculation during min/max freq read and enhance stats activation
When reading the min/max frequency, ensure that data calculations are
included. Without these calculations, some parts of the frequency data
may be lost.

Writing 0 disables the dump of min/max frequency statistic through the
interface and writing 1 enables it. Need to optimize the cleanup of the
min/max frequency statistic when writing 0 and restart statistic when
writing 1.

Additionally, treat the calculation as the caller of the cpufreq
transaction hook. This will update the time on the min/max frequency
across different governors.

Change-Id: I5989c8f9ce64546dad068cb5db944f2264d3c40c
Signed-off-by: Song Xue <quic_songxue@quicinc.com>