sched/walt: Introduce freq to cost tables
Currently, WALT sets up a per-cluster util-to-cost table of size 1024 to
precisely lookup the associated cost of any util when computing energy
values. An implicit assumption is made to boost the util by 25%,
representing a target load of 80. However, with the introduction of
zone-based target load in ("sched/walt: Leverage zone-based mechanism
for cpufreq"), this assumption may be incorrect.

Introduce a mechanism to map the util to a frequency, adjusted for the
proper target load. Drop the usage of util-to-cost tables and switch to
frequency-to-cost tables, indexed only for the necessary available
frequency levels.

Furthermore, create a mechanism to overwrite the v^2 * f based cost
tables with custom tables introduced at the SoC level. In the event of a
mismatch between the number of levels prescribed in the custom tables
and the available tables on the device, default to using the v^2 * f
based cost tables.

Change-Id: I352825add2d932a9d6372c3459d510282d8d6ebe
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>