leds: flash: leds-qcom-flash: ignore disable torch LED when reenable it
Commit 91fe1bda4672 ("leds: flash: leds-qcom-flash: Disable LED when
changing brightness") made a change to de-strobe a flash LED channel
and disable it whenever the brightness is changed for torch mode.
Even though this is functional to operate torch LED like this, it causes
a bad user experience when the torch LED is required to be reenabled,
such as when zooming in/out camera with other camera sensors with torch
LED enabled while it is already turned on by a camera sensor, a flicker
in brightness will be seen.

Fix this by not allowing to disable the LED channel if it's already
enabled in torch mode i.e. allow changing the brightness level for
torch mode.

Change-Id: I520ee072e98ee8bf582ac37c0f5e35340a60a01d
Signed-off-by: Fenglin Wu <quic_fenglinw@quicinc.com>