sched/walt: Fix WALT PD cost calculations during EAS
From 6.6 kernel onwards, to improve runtime of cost calculation in EAS
hotpath, the cost of a performance state (ps) i, in a performance domain
(pd), is calculated as:
  cost_i = (power_i * 10) / performance_i
where performance_i is calculated by considering a linear relationship
with frequency as:
  performance_i = (max_capacity * freq_i ) / fmax_pd

This prevents division by the cpu capacity (scale_cpu) while calculating
the cost.  Thus we can now simply calculate cost associated with a util
as:
pd->cost * \Sum cpu_util

This requires a fix in the WALT energy calculation as well. Currently,
we are dividing the cost by scale_cpu, which results in an incorrect
cost calculation due to an extra division. Removing this division is
necessary as it is not required.

Change-Id: I335e397ab380eb91bb4a0792f1bca49077124267
Signed-off-by: Atul Pant <quic_atulpant@quicinc.com>