soc: qcom: smp2p: Reset entry last value on SSR
When using smp2p version 2, the remote does not wait for the ssr ack
before setting the items. To accommodate this, set the last_value of all
the entries to 0 when SSR is detected. This forces smp2p to detect the
new values written by the remote. Because the SSR ack is skipped, the
down transition of bits is missed in smp2p version 2.

Change-Id: I54ea9110664a3d2891e90ecb410cc1428ca508d7
Signed-off-by: Chris Lew <quic_clew@quicinc.com>