minidump: Replace direct use of debugfs_list_mutex and debugfs_list
To be able to use debugfs_list and debugfs_list_mutex in the minidump
module, it needs to be exported from the kernel and added to the symbol
list.

There was a comment upstream that we would need to change the name of
these symbols before we can add them to the symbol list because their
names were too generic.

Hence I assume it might be better to use an exported API to retrieve the
dmabuf debugfs data rather than changing the name of the symbols in ack.

Change-Id: I4b354579462b33252af05131bd33b6965181dd65
Signed-off-by: Oreoluwa Babatunde <quic_obabatun@quicinc.com>