input: pm8941-pwrkey: Add DS and Hibernate support for pwrkey
On Hibernate entry, PMIC H/W context is lost and S/W context
is preserved. To make sure that H/W and S/W context are in sync
free the IRQs in freeze callback and in restore callback re-initialize
the H/W and re-register the IRQs to make sure that S/W and H/W context
are in sync. Similar change is required for Deep-Sleep use-case as well.

Add changes to support Deep-Sleep and Hibernate feature for pwrkey
driver.

Also,Add pull-up and delay changes to the pwr-key structure
and add config CONFIG_DEEPSLEEP to avoid breaking of
targets which do not support deep sleep .

Change-Id: I12abfb6a864f16e23676c2cd822ef804f6f0b952
Signed-off-by: Umang Chheda <quic_uchheda@quicinc.com>
Signed-off-by: Ayyagari Ushasreevalli <quic_aushasre@quicinc.com>