usb: dwc3: dwc3-msm-core: Eliminate redundant SSPhy power-up on device add
In the current implementation, the function dwc3_msm_host_ss_powerup()
is invoked for SS and SSP peripherals. This is redundant because the
host initialization process from the core driver already includes the
necessary steps to call ssphy_init, which also involves activating the
pipe_clk. The redundant attempt to switch on pipe_clk causes a glitch,
leading to multiple re-enumerations of any connected SS/SSP devices
before they are ready for operation.

Fix this issue by removing the call to dwc3_msm_host_ss_powerup()
during the SS/SSP device add event when a peripheral is connected.

Change-Id: I431852b4bd6f56d035006dc98ee79cfae81dc985
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>
Signed-off-by: Elson Roy Serrao <quic_eserrao@quicinc.com>