membuf: Avoid double-invoke of destructor callbacks
If a dma-buf has a client destructor notification, then prior to this
patch it would be invoked twice - once when the local dma-buf is closed,
and again when the gunyah MEM_RELEASED notification is handled. Fix this
by setting the destructor to null after it succeeds.

Also, if MEM_RELEASED is received before the local dma-buf is closed,
ignore it. Reclaiming the memory will occur when the local dma-buf is
closed instead to avoid changing state behind the dma-buf client's back.

Change-Id: I3504aac5e47ccb31b1ccdc6a6fe0c1ad24cd37b6
Fixes: 96c252b8f90b ("mem-buf: Listen for release notifications")
Signed-off-by: Patrick Daly <quic_pdaly@quicinc.com>
Signed-off-by: Vivekanand Tryambake <quic_vtryamba@quicinc.com>