phy: qualcomm: Dont assert reset upon QMP Combo usb exit
The expectation with asserting the reset during usb phy exit
is that the DP driver would toggle the assert during DP init
operation. But since DP Phy Ops are currently unused in the
combo driver, the resets would be left in the asserted state
when we power down phy in 4 lane DP mode. This would break
the DP functionality. Avoid this by skipping the assert
during phy exit.

Change-Id: If84399d77f2ded5640a629de70693fe858313322
Signed-off-by: Elson Roy Serrao <quic_eserrao@quicinc.com>