sched: walt: cached wallclock for colocation tick
Currently, the code is reading the hw clock every tick (on every cpu)
only to check if it needs to calculate colocation load.

This reading of hw clock, while is precise, but is unnecessary for
such a purpose. Instead use the cached clock the code path has
read recently.

For the wakeup path, the clock is read at the entry of the function.
While for the tick path, use rq->clock, it should have been read right
before the tick tracehook gets called.

Change-Id: I22b4b948a03664627e2e906aa193ad8eab92b832
Signed-off-by: Abhijeet Dharmapurikar <quic_adharmap@quicinc.com>