sched/walt: Fix issue with pipeline task being misfit
Currently, a task with a pipeline CPU of prime is always treated as a
misfit task, thereby always being treated as "prime worthy". However,
under nr=2, a pipeline task with a prime CPU should only be treated as
misfit if it is truly misfit.

Therefore, rework pipeline_Fits_smaller_cpus to ensure that tasks with
prime pipeline CPUs are treated as misfit only under nr=3.

Change-Id: I7ad058b4c47e3dd7d065a103197dcc204f531e82
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>