sched/walt: Fix issue with pipeline task being misfit
Currently, a task with a pipeline CPU of prime is always treated as a
misfit task, thereby always being treated as "prime worthy". However,
under nr=2, a pipeline task with a prime CPU should only be treated as
misfit if it is truly misfit.

Therefore, rework pipeline_Fits_smaller_cpus to ensure that tasks with
prime pipeline CPUs are treated as misfit only under nr=3.

Change-Id: I1e0540893328717000682a63de9d02f758433f09
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>