sched/walt: Add affinity criteria for pipeline tasks
If a task has an affinity mask that does not include any of the pipeline
CPUs, there can be issues with assigning pipeline CPUs. This can happen
as the task would be identified as pipeline, assigned a pipeline CPU,
but ultimately, the task would not be able to run on that CPU. This can
cause significant regressions as the task is consuming a valuable
pipeline CPU slot, while not even being able to exercise the benefits of
pipeline.

Therefore, ensure that any task vying to become a pipeline task is
eligible to be placed on any of the pipeline CPUs.

Change-Id: I5485970af4d182288bd5c3422ce32d7bbc873a9c
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>