sched/walt: Introduce Pipeline Energy Aware Scheduling
When three pipeline tasks are identified under the AUTO pipeline,
modify the pipeline behavior to assign tasks to pipeline CPUs with
energy awareness. During each pipeline placement evaluation, which
occurs at a 100ms cadence, evaluate the energy of two potential
configurations and select the one with the lowest energy.

Configuration 1 (1 Large):
Only the largest of the three tasks is placed on a
Prime CPU, while the other two tasks are placed on Gold CPUs.

Configuration 2 (2 Large):
The smallest of the three tasks is placed on one of
the Gold CPUs, while the other two tasks are placed on Prime CPUs.

Address IPC degradation issue by implementing an IPC degradation factor
to reduce IPC when multiple tasks run on the prime cluster.
Ensure that when a second task is placed on a large cluster, the IPC
benefit of the first task is degraded due to shared cache usage,
ensuring that the energy evaluation of prime factors in the
resulting IPC degradation.

Finally, introduce a 2L scaling factor, which acts as an additional
correction on the prime energy, when evaluated under 2L.

Change-Id: I74ea3ac37627a22131fb63875bb73fc42702fd41
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>