gunyah: support polling vcpu run
When there is a kernel panic event on PVM, give a
chance to SVM and SVM can handle the panic notify
from PVM side. This notify is one doorbell and
doesn't know the interrupt in which vcpu. So provide
one API to poll vcpu run and make SVM can handle
this interrput.

Change-Id: I78b25ac080ba8a40021a7a12df943ebb6dcf12e9
Signed-off-by: Peng Yang <quic_penyan@quicinc.com>