sched: walt: quick return from max load reports
Currently the freq updating code walks through various
situations and chooses to either amplify the load or
force an fmax worthy load, not in any order.

This ends up with wasted cycles where other situations
are checked when there is an upcoming situation that forces
fmax.

Optimize the code flow such that once fmax is forced other
situations are skipped.

Change-Id: Ie8a4c5109b00e5d414b459155a5b87d4a74d2c78
Signed-off-by: Abhijeet Dharmapurikar <quic_adharmap@quicinc.com>