sched/walt: Introduce pipeline demand scaling
Pipeline demand viewed on gold cluster is not equivalent when view from
prime cluster.

Introduce a default scaling factor of 70. This scaling factor acts as
an IPC ratio fixup, ensuring that the task utilization of a pipeline
task running on Gold CPUs is not read as inflated when evaluating
moving it to Prime CPUs. Similarly, the utilization of a task running
on Prime CPUs should appear boosted when evaluating what its utilization
would be if it were running on Gold CPUs. For example, if a 200
utilization task on Gold CPUs should appear as 140 utilization on Prime
CPUs, and vice versa, a scaling factor of 70% would help achieve this
fixup for energy evaluation purposes.

Without this fixup, it is incorrect to compare demand values of tasks
running on golds and primes as they are not one-to-one. To better
capture this difference, ensure that pipeline_demand always returns two
values: the gold equivalent demand and the prime equivalent demand.

By comparing scaled demand values, pipeline selection is more smooth and
there is reduced ping-ponging of the pipeline task selection.

Change-Id: Ib59c1ab1a5d13a7be6424b775e2134e1b74a920b
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>