sound: usb: usb_audio_qmi_svc: Update error path code
check_valid_request() was introduced to split apart some of the USB stream
enable/disable request checks, in order to determine if the command
received from the aDSP is valid.  Some of the error paths were not properly
updated to account for the code refactoring.  Updating the function to
handle find_substream() errors properly.

Change-Id: Ic31dcd52de2d8ccf71ea62867dc07b25ed138d49
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>