thermal: qcom: Refine tsens reading logic to ensure valid read
The current tsens reading logic waits for the valid bit
to become 1 before reading the temperature. However, this
approach can result in invalid reads when the tsens reading
is performed at very short intervals. A corner case exists
where the tsens driver might read the register while the
tsens hardware is updating the register, potentially leading
to an invalid value being read. Therefore, tsens readings and
the valid bit should be read simultaneously, and the reading
should be considered valid only if the valid bit is 1.

This update modifies the tsens sensor reading logic to ensure
it aligns with the hardware recommendation.

Change-Id: I6cec1cbea4e7dec2950afd76435b6da4fcbb964b
Signed-off-by: Priyansh Jain <quic_priyjain@quicinc.com>