input: qcom-hv-haptics: update set_gain() before play is enabled
The API set_gain() would get called before the play source is determined
which is usually happened when the play is triggered. However, in such
case, it is only useful for DIRECT_PLAY mode which has an independent
register to specify the amplitude. Therefore, set the DIRECT_PLAY
amplitude when set_gain() is called before haptics play is triggered.

Change-Id: I741195db30e112881747739c615fb55b019507dd
Signed-off-by: Fenglin Wu <quic_fenglinw@quicinc.com>