sched/walt: Reduce excess walt dump prints
When WALT dumps up/down migration values, it is currently dumping for
all CPUs, despite the fact that the values will all be the same within
the same cluster. This is unnecessary.

Furthermore, the prints aren't very clear. Modify this to clearly print
the specific cgroup associated with each value.

Change-Id: If142221b4c8ab9f53427fc5cb02ca5d38bf1e2ee
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>