clk: qcom: videocc-canoe: Update videocc spare1 register
On canoe v1 platform, mvs0c clk divider was updated to run mvs0c_clk
at the same frequency as mvs0_clk, to avoid random video hang issues
due to a HW issue. But this HW issue is fixed on canoe v2 platform, and
hence the v1 workaround is not required on canoe v2 platform. But to
enable the HW fix, SW need to set bit 0 in videocc spare1 register.
Hence set bit0 in videocc spare1 register as per HW recommendation on
canoe v2 platform.

Setting this chicken bit on v1 will not do anything so write this
irrespective of canoe versions.

Change-Id: I300512ce50aff1cb9a6a7f2daf9720164031d6b1
Signed-off-by: Vivek Aknurwar <quic_viveka@quicinc.com>