usb: typec: ucsi_qti_glink: Add poll_cci callback function
Commit 976e7e9bdc77 ("acpi: typec: ucsi: Introduce a ->poll_cci method")
added a new ops poll_cci() in the ucsi framework. This had caused
the driver registration with the framework to fail runtime.

Add poll_cci callback function to align with the new UCSI framework
check for this callback. Keep the function used the same as the one
for read_cci callback function as it is sufficient for the present
requirements.

Change-Id: Ifc796ec527f5465f6c47697dd9e152a128093877
Signed-off-by: Jishnu Prakash <quic_jprakash@quicinc.com>