zram: qpace: Remove scheduling bubbles
Even when we attempt to queue up another compression item inside
zram_qpace_work_fn(), with the goal of being rescheduled immediately,
this will not always happen. Thus, so long as work is queued up, let's
not exit zram_qpace_work_fn() but instead head back to the start.

Whilst we busy-loop for copying to finish, 100 us is the minimum
required in udelay() to use more efficient polling, but copying data
will finish much earlier when using a 508 KB submission size.  Let's
reduce the udelay() time to just 1 us accordingly.

Replace our workqueue with a kthread. We do this because our backing
worker pool is shared with other workqueues, which can lead to two
problems: (1) workqueues have a timeout for how long a given work item
can run for - without a limit, other items would be starved, which
will then trigger a so called workqueue lockup. (2) Conversely, we
can get delayed waiting for a worker to be assigned to us, even if we
queue up work.

Change-Id: I69711e4a42f1368b44205701d5925b2a9d85ec33
Signed-off-by: Chris Goldsworthy <quic_cgoldswo@quicinc.com>