drivers: tlmm_vm_mem_access: Don't read rm-mem-handle from the DT
RM will not be able to patch the tlmm-vm-access node. So, the driver
won't be able to get the appropriate mem_handle to issue the RELEASE
notification. Let HLOS initiate a SHARED notification when OS status is
set because the HLOS knows what the mem-handle is and the VM can then
use this mem-handle to initiate the RELEASE notification.

Change-Id: Ib77a06780172db9d0ebf2421784da23db27bf4d3
Signed-off-by: Prakruthi Deepak Heragu <quic_pheragu@quicinc.com>
Signed-off-by: Murali Nalajala <quic_mnalajal@quicinc.com>