sched: walt: enhance busy time bugs
Currently code serially subtracts busy time trackers and issues a bug
when it encounters a tracker with -ve value. This makes debugging a bit
complicated as it is hard to decipher which trackers have been
successfully subtracted and which ones aren't.

So check all the trackers for -ve result and commit the subtraction only
after all the trackers are guaranteed to be +ve.

While at it , create and use a simple macro to do a clamped subtraction
i.e. return 0 if the subtraction result is -ve.

Change-Id: I1b46ff50786b5205e02ac1adc381b86723e0a6b7
Signed-off-by: Abhijeet Dharmapurikar <quic_adharmap@quicinc.com>