qpace: zram: Don't use hw_write_ptr in certain functions
Inside of qpace_free_tr_entries(), we use hw_write_ptr to free up
entries we've already consumed, on the assumption that hw_write_ptr
hasn't changed after we call qpace_trigger_tr(). Note that if we call
qpace_trigger_tr(), and a user comes and queues additional pages in
QPaCE, hw_write_ptr will be incremented, which will break
qpace_free_tr_entries(), as we will end up freeing / ignoring requests
that were submitted to us.

Additionally, simplify transfer_ring_is_empty() by just checking for
the number of items.  For this to work, we must, modify get_qpace() to
not be tied to any specific ring, that will incorrectly lead to there
being an item queued when that is not the case, which would
break transfer_ring_is_empty() for the ZRAM use case, as we call
get_qpace(COMPRESS_RING) today.

Change-Id: Ie7a42c7b01b7c5de57dc344f5eed7845a1bd3135
Signed-off-by: Chris Goldsworthy <quic_cgoldswo@quicinc.com>