clk: qcom: dispcc-canoe: Move dispcc MX clocks to separate device
Few clocks on dispcc are on MX rail, which is sufficient for these
clocks to operate and there is no need for MMCX rail to be ON. But
the current dispcc driver votes on MMCX rail when any dispcc clock
is enabled as part of its runtime resume callback. Move the dispcc
MX clocks to a separate clock controller device to avoid voting on
MMCX when these clocks are enabled.

However MMCX is still required to access these clock registers, so
client drivers should ensure there is a vote on MMCX rail before
calling any clock API on these dispcc MX clocks and remove the
MMCX vote after that.

Change-Id: Ifc2c77e3f09882574522a1f50c72a83ba0c5628f
Signed-off-by: Jagadeesh Kona <quic_jkona@quicinc.com>