sched: walt: allow LRPB continuous boost
The current design of the LRPB is to boost frequency for a specific
period (based on the window size). After this boost period, the
system load is expected to drive the frequency to the desired level.

However, with the introduction of the smart frequency framework and the
LRPB override, the previous frequency guidance is no longer valid. As
soon as the LRPB period ends, the frequency is capped again by the smart
frequency framework.

In scenarios where the smart frequency framework is primarily used for
capping (with most reasons disabled) to save power, and the system
relies on LRPB to boost frequency, a temporary boost is insufficient.
Therefore, it is necessary to allow LRPB to boost the frequency for as
long as it is active.

Change-Id: I788a8b705f76d4c0f49aabfb80fa5efd79422336
Signed-off-by: Ashay Jaiswal <quic_ashayj@quicinc.com>