pinctrl: qcom: Add user modification of the TLMM dir_conn on sa8797p
Add support for user modification of the dir_conn list in order to
support use cases where different GPIOs need to be routed as interrupts
via the direct connect IRQs on the sa8797p.

This snapshot is taken as of qcom-6.1 branch commit 3bf1bccb1912d
("pinctrl: qcom: Add user modification of the TLMM dir_conn
on sa8797p").

Change-Id: I6f61a046162984f919a24b6f0710fcb4b9063c75
Signed-off-by: Siddharth Patil <quic_siddpati@quicinc.com>