mem-buf: Prevent use-after-free by dma-buf client destructor callbacks
Clients are allowed to register a callback & private data,  which gets
invoked during dmabuf->ops->release(). In commit 5fd3e2893c55 ("mem_buf:
Acquire/Release an extra reference on state transition") a more complex
release model was added which can retry the release process multiple
times. If the client's private data contains data which is kfree'd on
the first release attempt, a use-after-free will occur on subsequent
calls. In particular, passing the dma-buf as the private data causes
this.

Revert the client destructor to the prior model, in which it is called
only once, and add comments indicating the return value from the
client destructor is ignored.

Change-Id: Ic79cb6d550bc7c39dbb3d35331a29683d99ad4de
Signed-off-by: Patrick Daly <quic_pdaly@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>