sched: walt: don't reset legacy smart_freq for malformed input
Currently, if a malformed input is given to the legacy_smart_fmax node,
it resets the existing valid settings. Do not do that.

While at it, delay grabbing the mutex freq_reason_mutex until the input
has been confirmed to be good. Also change how the return value is
tracked, i.e. assume success unless an issue is encounted, upon which
set it to -EINVAL and bail.

Change-Id: I34ea5397506be12ab8297b54d684e2676a3c808a
Signed-off-by: Abhijeet Dharmapurikar <quic_adharmap@quicinc.com>