sched: walt: fix capacity order of smart freq
Currently, each cluster independently change its smart_freq cap (either
the legacy or the ipc one) which results in changing its capacities.

With this the capacity order between cluster can break when
a higher cluster remains capped at a lower freq while a lower cluster
gets uncapped.

This problem can be easily solved if uncap freq is imposed on all the
higher clusters. The higher ipc of the higher clusters ensures the same
freq results in a higher capacity.

The fix for this needs to be handled differently for legacy and ipc
smart freq.
For legacy smart freq scheme, since smart freq cap is determined
sequentially in increasing order of cluster capacities, consider the uncap
freq from the previous cluster.

For ipc smart freq however, we update the cpufreq of all the higher
capacity clusters sequentially upon which they get an opportunity to
peek in to the earlier cluster's uncap freq and adjust their own.

While at it we also ensure that if the freq are uncapped to
FREQ_QOS_MAX_DEFAULT_VALUE we use the cluster's max_possible_freq
to set the higher cluster uncap level.

Note that walt_internal_freq_limit of the prev cluster could change
right after its consumed for capacity checks. This is not a problem
for legacy smart freq because the limit is changed by just one cpu that
does the window rollover.
For IPC multiple cpus from a lower
cluster could cause a change to walt_internal_freq_limit. However
the last cpu triggering the WALT_CPUFREQ_SMART_FREQ_BIT gov
callback will observe the latest ipc updates from all the cpus
leading to an accurate evaluation for capacity ordering.

Change-Id: I12c71227df79eadb7926e372aefb315f9e1d116f
Signed-off-by: Abhijeet Dharmapurikar <quic_adharmap@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>