sched: walt: flag to disable capacity sync
Currently, if the internal limit of a lower cluster changes, the upper
cluster internal limit is at least brought up to the level of the prev
cluster.

While this is desired all the time, certain power sensitive applications
need to keep the higher clusters suppressed.

Provide means to enable/disable capacity sync using a feature bit.

Change-Id: I9a8f69603ed292b1750c14a9e18377a14500fc68
Signed-off-by: Abhijeet Dharmapurikar <quic_adharmap@quicinc.com>
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>