sched: walt: flag to disable capacity sync
Currently, if the internal limit of a lower cluster changes, the upper
cluster internal limit is at least brought up to the level of the prev
cluster.

While this is desired all the time, certain power sensitive applications
need to keep the higher clusters suppressed.

Provide means to enable/disable capacity sync using a feature bit.

Change-Id: I139665ac752b99d2c98d99e945025304cbc3abea
Signed-off-by: Abhijeet Dharmapurikar <quic_adharmap@quicinc.com>
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>