mem-buf-gh: fix compile error for !CONFIG_DMA_CMA
When CONFIG_DMA_CMA (or CONFIG_CMA) is not enabled, the struct device
will not have the struct cma *cma_area field, leading it to compile
errors since cma_area is used in the code.

Wrap the cma_area related code with CONFIG_DMA_CMA so that they
get compiled out if the confid is disabled.

In the case when CONFIG_CMA is disabled, it automatically disables
CONFIG_DMA_CMA as well, so this case is covered as well.

Change-Id: I0ebc4da7223105c2cc8008d307575b33dfb27351
Signed-off-by: Sudarshan Rajagopalan <quic_sudaraja@quicinc.com>