sched/walt: Remove unsafe walt_map_util_freq usage
The function walt_map_util_freq was used to translate util to
frequency based on zone util thresholds and determine the correct target
load. However, this relies on wg_policy being active, which is only true
when the WALT governor is in use.

If WALT is not the active governor, wg_policy will be freed, leading to
a crash. Temporarily remove the problematic invocation and manually
scale the util, assuming a fixed target load of 80, which translates to
a 25% boost. This avoids the dependency on wg_policy and prevents
any instability.

Change-Id: I552e2a0c200d8962d25a7f831e24bf9f8bd7072b
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>