serial: msm_geni_serial: Handle runtime PM when tty port is closed
A NOC error is observed during the terminate_all callback in the
cancel_tx function. This occurs because the port is already closed
before the termination sequence is executed, leading to hardware access
violations.

Add a check to verify if the port is open before invoking the
termination sequence. If the port is closed, ensure that necessary
resources are enabled using pm_runtime_get_sync() to avoid NOC errors.

Change-Id: I00435cfca1e5f9c4327fd5d1e02bb5017029e62a
Signed-off-by: Somesh Dey <quic_somedey@quicinc.com>
Signed-off-by: Viken Dadhaniya <viken.dadhaniya@oss.qualcomm.com>