sched/walt: Fix corruption issue in tick tracehook
Runqueue information is passed to the tracehook associated with
scheduler tick via the rq variable. This information is vital as it tells
which CPU tick is occurring on. This rq variable is overwritten by code
that traverses all the clusters and assigns the cluster's first CPU's rq
to it.
Fix this by creating another variable to traverse the clusters'
runqueues, thus preserving the rq that was passed to the tracehook.

Change-Id: Ia60d6f898385aa0901abcdc2e95b23cd4d769ccf
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>