sched/walt: update PIPELINE fastpath evaluation
Current implementation of PIPELINE fastpath uses prev_cpu
(previous cpu) as backup pipeline cpu if the target pipeline
cpu is already running a pipeline task, this is done based on
expectation that prev_cpu is always a pipeline cpu for any
pipeline task.
There are cases where prev_cpu might not be a pipeline cpu
(one such case is when a task gets selected as pipeline task
for first time and running on a non-pipeline cpu) in this
situation prev_cpu of task will be a non-pipeline cpu and
WALT will use PIPELINE_FASTPATH with a non-pipeline cpu.

Add check to select prev_cpu only if it is a pipeline cpu,
otherwise find a free pipeline cpu and place pipeline task.

Change-Id: If195edae3d9db7c78497aaa8448d42505fa171c9
Signed-off-by: Ashay Jaiswal <quic_ashayj@quicinc.com>