zram: Delay setting comp_len on account of replicated word branch
If we find that a page is composed of replicated words, we want to
make sure that we do not pass PAGE_SIZE to zram_write_finish(), as it
would mark a page as both being a huge page and a replicated word
page. This issue has not been seen, but to avoid any future changes in
QPaCE's behavior, pass in a comp_len of 0 to zram_write_finish() if we
see that a page is composed of replicated bytes.

Change-Id: Ie730f3221efaa54fe4a7afe7a1e19cea718cb5a6
Signed-off-by: Chris Goldsworthy <quic_cgoldswo@quicinc.com>