bus: mhi: host: Fix firmware load image pointer
The tme_supported_image support was added for the use case where
request_firmware() is used to initialize a struct firmware. Upstream mhi
code supports an alternate method where the mhi controller driver
provides a direct reference to the firmware itself, bypassing the use of
request_firmware().

The issue was identified with static analysis tools:
    "Null pointer 'firmware' that comes from line 419 may be
     dereferenced at line 521."
419         const struct firmware *firmware = NULL;
...
520                 if (mhi_cntrl->tme_supported_image) {
521                         fw_data = firmware->data + mhi_cntrl->sbl_size;
522                         fw_sz = fw_sz - mhi_cntrl->sbl_size;
523                 }

To address the missed case, use the fw_data pointer which will support
both methods of firmware loading.

Change-Id: I2aa242d766ba98aeda5fd1526c9db681e307afe5
Signed-off-by: Matthew Leung <quic_mattleun@quicinc.com>