pci: msm: Remove dev check in pm functions
In the pm resume and suspend functions, the RC device reference is
initialized from private data of the pci device's bus. The
initialization implies that the pci device is valid making the extra
guard check unnecessary. This error was identified with static analysis
tools.

To address this error, remove the unnecessary pci device check.

Change-Id: I50f4f493d51119c59d581354398c5c2aa8cbaa0a
Signed-off-by: Matthew Leung <quic_mattleun@quicinc.com>