iommu: Move qcom_iommu_group_parse_phandle() to be called after NULL check
The 'dev' variable is passed as an argument to the
qcom_iommu_group_parse_phandle(). This happens right before 'dev' is
checked for NULL.

'dev' should be checked for NULL first before being used anywhere within
the function to verify that the variable is non-zero. Hence, move the
qcom_iommu_group_parse_phandle to be below the NULL check.

Change-Id: I9a5625c6b7041d5aa71a75c12240da412b8f812e
Signed-off-by: Oreoluwa Babatunde <quic_obabatun@quicinc.com>