minidump: Restructure how the offset value is tracked for slabowner data
The offset value helps keep track of the free buffer space in the memory
allocated to store the dump data.

Restructure how the offset value is updated to make it cleaner.
Also add a check at the start of the dump_tracking() function to make
sure that this value never exceeds the buffer size to prevent overflow.

Change-Id: I558369553d4795aadaf5879290a59ba84ada7bed
Signed-off-by: Oreoluwa Babatunde <quic_obabatun@quicinc.com>