sched/walt: Fix core ctl enablement under pipeline
When the pipeline operates with nr=2, core control boosting should be
disabled.

Support for single-thread pipelines was introduced in commit
db594dccfade ("sched: walt: add support for single thread pipeline").
However, the logic for pipeline unisolation was incorrectly implemented
to trigger whenever a prime CPU is present in the sched_pipeline_cpus
mask.

This causes incorrect behavior in nr=2 scenarios where a prime CPU is
included in the mask. To fix this, explicitly check for the
single-thread pipeline use case before deciding to unisolate cores.

Change-Id: I900c4415a8aeed8c9e954156e2c384e10ab7acb2
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>