soc: qcom: fsa4480-i2c: Register Type-C Mux device
Add Type-C Mux device to fsa4480 driver so that it can receive
events from Type-C framework for the Type-C mode from which
it can configure the settings for audio accessory or otherwise.

With this change, there is no need to register for ucsi_glink
notifier to receive the partner accessory information. Hence
remove the code related to that.

Change-Id: I9817dfbe5f20996b5ae1e47e478da3e2ea9a9744
Signed-off-by: Ramireddy KrishnaKanth Reddy <quic_ramikris@quicinc.com>