qpace: Replace individual urgent register writes with stp
Cut down on the time it takes to trigger urgent command writes by
replacing individual 32-bit register writes for triggering an urgent
command with an stp ("store pair") instruction:

	stp <lower-reg>, <upper-reg>, [<address>]

In AArch64, this would be internally be equivalent two 64-bit
stores with

	mem[<addr>,<addr> + 7] = <lower-reg>
	mem[<addr> + 8, <addr> + 15] = <upper-reg>

executed in that order. With our code, we would write to the same
registers in the same order as before, but we would combine the two
output-register-writes into a single write, and the two
input-register-writes into a single write.

Change-Id: Ia14c70564bf9a7525d252ec224a2f017d89f5567
Signed-off-by: Chris Goldsworthy <quic_cgoldswo@quicinc.com>