sched/walt: Fix pipeline demand eval for displaced tasks
When selecting for the top 3 pipeline tasks, the group list is
traversed. However, this list is not sorted by demand, which is the
primary condition used to evaluate which tasks will be considered for
pipeline.

During evaluation, a task already placed in heavy_wts may be displaced
by one with higher demand. When this occurs, the displaced task is
reconsidered for other available slots by being repurposed as the new
to_be_placed_wts.

Previously, this re-evaluation used stale demand values during swaps.
Ensure that current demand values are used when reassessing displaced
tasks.

Change-Id: Iac89629e381fee59de42570377e382214d957982
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>