drivers: dcvs: memlat: Remove cpucp sample ms
Remove the cpucp_sample_ms sysfs parameter as it is not
used today. All clients are instead using the legacy
sample_ms parameter which updates the setting for both
hlos and cpucp versions of  memlat. This will also help
avoid any confusion from having two separate parameters
as there is no need today for them to be different.

Change-Id: Ifa8b8e471f64e7bf7f79408ae4e571c22336adc3
Signed-off-by: Amir Vajid <quic_avajid@quicinc.com>