sched/walt: Introduce trailblazer_boost_state
Introduce a new metric, trailblazer_boost_state, which activates when a
trailblazer task has been active on primes and the corresponding IPC of
the prime CPUs is or exceeds 300. In addition, add a 100ms hysteresis to
prevent rapid toggling.

When trailblazer_boost_state is active:

- DDR boost is triggered to fmax
- The floor frequency for gold cores is raised from 1GHz to 2.5GHz,
gated by the same IPC threshold.

The trailblazer_boost_state mechanism is independent of the existing
trailblazer_state, which is based on a 40% activity window. For DDR
boost, trailblazer_state is not used, however, for CPU Frequency floor
frequency elevation, the trailblazer_state continues to be used.

Change-Id: Ia789b2350477c30a1d297d5b6b1fc6627a27f81a
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>